+++ kompendium-locations/src/test/kotlin/io/bkbn/kompendium/locations/util/TestResponseInfo.kt
@@ -1,12 +1,16 @@
 package io.bkbn.kompendium.locations.util
 
-import io.bkbn.kompendium.core.metadata.MethodInfo
+import io.bkbn.kompendium.core.metadata.method.MethodInfo
 import io.bkbn.kompendium.core.metadata.RequestInfo
 import io.bkbn.kompendium.core.metadata.ResponseInfo
+import io.bkbn.kompendium.core.metadata.method.DeleteInfo
+import io.bkbn.kompendium.core.metadata.method.GetInfo
+import io.bkbn.kompendium.core.metadata.method.PostInfo
+import io.bkbn.kompendium.core.metadata.method.PutInfo
 import io.ktor.http.HttpStatusCode
 
 object TestResponseInfo {
-  val testGetSimpleLocation = MethodInfo.GetInfo<SimpleLoc, SimpleResponse>(
+  val testGetSimpleLocation = GetInfo<SimpleLoc, SimpleResponse>(
     summary = "Location Test",
     description = "A cool test",
     responseInfo = ResponseInfo(
@@ -14,18 +18,7 @@ object TestResponseInfo {
       description = "A successful endeavor"
     )
   )
-  val testPostSimpleLocation = MethodInfo.PostInfo<SimpleLoc, SimpleRequest, SimpleResponse>(
-    summary = "Location Test",
-    description = "A cool test",
-    requestInfo = RequestInfo(
-      description = "Cool stuff"
-    ),
-    responseInfo = ResponseInfo(
-      status = HttpStatusCode.OK,
-      description = "A successful endeavor"
-    )
-  )
-  val testPutSimpleLocation = MethodInfo.PutInfo<SimpleLoc, SimpleRequest, SimpleResponse>(
+  val testPostSimpleLocation = PostInfo<SimpleLoc, SimpleRequest, SimpleResponse>(
     summary = "Location Test",
     description = "A cool test",
     requestInfo = RequestInfo(
@@ -36,23 +29,7 @@ object TestResponseInfo {
       description = "A successful endeavor"
     )
   )
-  val testDeleteSimpleLocation = MethodInfo.DeleteInfo<SimpleLoc, SimpleResponse>(
-    summary = "Location Test",
-    description = "A cool test",
-    responseInfo = ResponseInfo(
-      status = HttpStatusCode.OK,
-      description = "A successful endeavor"
-    )
-  )
-  val testGetNestedLocation = MethodInfo.GetInfo<SimpleLoc.NestedLoc, SimpleResponse>(
-    summary = "Location Test",
-    description = "A cool test",
-    responseInfo = ResponseInfo(
-      status = HttpStatusCode.OK,
-      description = "A successful endeavor"
-    )
-  )
-  val testPostNestedLocation = MethodInfo.PostInfo<SimpleLoc.NestedLoc, SimpleRequest, SimpleResponse>(
+  val testPutSimpleLocation = PutInfo<SimpleLoc, SimpleRequest, SimpleResponse>(
     summary = "Location Test",
     description = "A cool test",
     requestInfo = RequestInfo(
@@ -63,7 +40,23 @@ object TestResponseInfo {
       description = "A successful endeavor"
     )
   )
-  val testPutNestedLocation = MethodInfo.PutInfo<SimpleLoc.NestedLoc, SimpleRequest, SimpleResponse>(
+  val testDeleteSimpleLocation = DeleteInfo<SimpleLoc, SimpleResponse>(
+    summary = "Location Test",
+    description = "A cool test",
+    responseInfo = ResponseInfo(
+      status = HttpStatusCode.OK,
+      description = "A successful endeavor"
+    )
+  )
+  val testGetNestedLocation = GetInfo<SimpleLoc.NestedLoc, SimpleResponse>(
+    summary = "Location Test",
+    description = "A cool test",
+    responseInfo = ResponseInfo(
+      status = HttpStatusCode.OK,
+      description = "A successful endeavor"
+    )
+  )
+  val testPostNestedLocation = PostInfo<SimpleLoc.NestedLoc, SimpleRequest, SimpleResponse>(
     summary = "Location Test",
     description = "A cool test",
     requestInfo = RequestInfo(
@@ -74,7 +67,18 @@ object TestResponseInfo {
       description = "A successful endeavor"
     )
   )
-  val testDeleteNestedLocation = MethodInfo.DeleteInfo<SimpleLoc.NestedLoc, SimpleResponse>(
+  val testPutNestedLocation = PutInfo<SimpleLoc.NestedLoc, SimpleRequest, SimpleResponse>(
+    summary = "Location Test",
+    description = "A cool test",
+    requestInfo = RequestInfo(
+      description = "Cool stuff"
+    ),
+    responseInfo = ResponseInfo(
+      status = HttpStatusCode.OK,
+      description = "A successful endeavor"
+    )
+  )
+  val testDeleteNestedLocation = DeleteInfo<SimpleLoc.NestedLoc, SimpleResponse>(
     summary = "Location Test",
     description = "A cool test",
     responseInfo = ResponseInfo(